+++ packages/heroku-pipelines/test/commands/pipelines/info.js
@@ -56,8 +56,8 @@ production-app-1   production`)
     pipeline = { name: 'example', id: '0123', owner }
 
     if (owner && owner.type === 'team') {
-      api.get('/teams/1234').reply(200, {
-        id: '1234',
+      api.get(`/teams/${owner.id}`).reply(200, {
+        id: owner.id,
         name: 'my-team'
       })
     }
@@ -100,7 +100,6 @@ production-app-1   production`)
 
   context(`when pipeline doesn't have an owner`, function () {
     beforeEach(function () {
-      pipeline = { name: 'example', id: '0123' }
       setup()
     })
 
@@ -120,24 +119,98 @@ production-app-1   production`)
   })
 
   context('when it has an owner', function () {
+    function itShowsMixedOwnershipWarning (owner) {
+      it('displays mixed ownership warning', function () {
+        return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
+          const warningMessage = ` ▸    Some apps in this pipeline do not belong to ${owner}.
+ ▸    \n ▸    All apps in a pipeline must have the same owner as the pipeline owner.
+ ▸    Transfer these apps or change the pipeline owner in pipeline settings.
+ ▸    See https://devcenter.heroku.com/articles/pipeline-ownership-transition
+ ▸    for more info.
+`
+          cli.stderr.should.contain(warningMessage)
+        }).then(() => api.done())
+      })
+    }
+
+    function itDoesNotShowMixedOwnershipWarning () {
+      it(`doesn't display mixed ownership warning`, function () {
+        return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
+          const warningMessage = `Some apps in this pipeline do not belong`
+          cli.stdout.should.not.contain(warningMessage)
+        }).then(() => api.done())
+      })
+    }
+
     context('and type is user', function () {
-      beforeEach(function () {
-        setup(setup({ id: '1234', type: 'user' }))
+      context('with mixed pipeline ownership ', function () {
+        // id '5678' means pipeline owner doesn't own any of the pipeline apps
+        const pipelineOwner = { id: '5678', type: 'user' }
+
+        beforeEach(function () {
+          setup(pipelineOwner)
+        })
+
+        it('shows uuid instead of email', function () {
+          return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
+            cli.stdout.should.contain('owner: 5678')
+          }).then(() => api.done())
+        })
+
+        itShowsMixedOwnershipWarning('5678')
+        itShowsPipelineApps()
+      })
+
+      context('with homogeneous ownership', function () {
+        const pipelineOwner = { id: '1234', type: 'user' }
+
+        beforeEach(function () {
+          setup(pipelineOwner)
+        })
+
+        it('displays the owner email', function () {
+          return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
+            cli.stdout.should.contain('owner: foo@user.com')
+          }).then(() => api.done())
+        })
+
+        itDoesNotShowMixedOwnershipWarning()
+        itShowsPipelineApps()
       })
     })
 
     context('and type is team', function () {
-      beforeEach(function () {
-        setup({ id: '1234', type: 'team' })
+      function itShowsTeamAsOwner () {
+        it('displays the owner', function () {
+          return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
+            cli.stdout.should.contain('owner: my-team (team)')
+          }).then(() => api.done())
+        })
+      }
+
+      context('with mixed pipeline ownership', function () {
+        const pipelineOwner = { id: '5678', type: 'team' }
+
+        beforeEach(function () {
+          setup(pipelineOwner)
+        })
+
+        itShowsTeamAsOwner()
+        itShowsPipelineApps()
+        itShowsMixedOwnershipWarning('my-team (team)')
       })
 
-      it('displays the owner', function () {
-        return cmd.run({ args: { pipeline: 'example' }, flags: {} }).then(() => {
-          cli.stdout.should.contain('owner: my-team (team)')
-        }).then(() => api.done())
-      })
+      context('with homogeneous ownership', function () {
+        const pipelineOwner = { id: '1234', type: 'team' }
 
-      itShowsPipelineApps()
+        beforeEach(function () {
+          setup(pipelineOwner)
+        })
+
+        itShowsTeamAsOwner()
+        itShowsPipelineApps()
+        itDoesNotShowMixedOwnershipWarning()
+      })
     })
   })
 })